refactor(core): 建立 admin 域骨架并切换挂载
